Responsive and fix profile
